Improve iocccsize_test.sh
Add -Z option to find top level directory (much like the other scripts
that have this option). This way one can run the script from test_ioccc
directly.

Add -I option to allow specifying of args to pass to iocccsize. Since
this is a variable it seems like it would be good if this can be set.
One of course would have to quote it. In usage string there is a note
that one should end it with --.

Added some exit codes. Instead of 10 >= being also for missing files
make those < 10 and make >= 10 for internal errors like the usage string
said.

When an invalid command line is issued the usage string is printed. This
was already there but now they're separated so it tells the user if the
option is unknown or if they need to specify an argument. Showing the
usage for invalid command line should probably be done for all the other
scripts too just like the compiled tool but for now they do not have it.

Updated the man page for these changes.