Minor cosmetic changes to dyn_array
Rename the macro DYN_ALLOC_VERSION as DYN_ARRAY_VERSION as that's what
the name of the facility is.

In the macro change the 'array_p' to be 'array' to match the functions.
This was a choice of simplicity: if we changed the \<array\> in the
files to be array_p it would also change places where we would not want
e.g. in the string 'dynamic array'. Thus since the functions used
'array' the macros now also use it.

Updated the man page for these changes including adding the version
macro to it (this needs to be done for the jparse man page too possibly
but that also needs a section 3 man page which might be where that
actually belongs).