Move all_sem_ref.8 to soup/man/man8
This is because the script is actually under soup/ not test_ioccc/ which
is where it was before (that is the man page was under
test_ioccc/man/man8 but the script is in soup/).

At the same time I changed the word 'trailer' to 'footer' as that's what
it actually is - a footer file (though it's not perfect it goes well
with the meaning of 'header' used).