Exit code corrections in run_usage.sh
Some of the exit codes were not used. Removed them from both the man
page and the script. As well the man page was incomplete in the way of
exit codes. See below for a list of what files/tools still have to be
checked for exit code consistency (including with seqcexit comments).

For the man page I also added a note about how it's a quick, dirty and
ugly hack and that it will soon be deprecated (though it has one use
still at that point).

The following tools still need to be checked for exit codes along with
their respective man pages:

    jsemcgen.sh
    jnum_chk
    run_bison
    jsemtblgen
    jnum_gen
    verge
    jsemcgen
    jstr_test
    run_flex
    prep
    mkiocccentry_test
    vermod
    ioccc_test
    hostchk
    jparse_test
    txzchk_test
    utf8_test
    chkentry_test
    reset_tstamp
    bug_report
    jstrdecode
    chkentry
    jstrencode
    jparse
    fnamchk
    iocccsize
    txzchk
    mkiocccentry