Add bug_report.sh.1 man page (symlink)
As the man page is for bug_report.sh it seems that having a
bug_report.sh.1 man page is also a good idea. There are some scripts
left that don't yet have a symlink in the same way but perhaps should
(or else the symlink should be removed and the file renamed).