Check exit codes for jsemcgen.sh
At the same time I have made some slight modifications to formatting and
wording. There are some things that must be figured out in this before
more can be done with checking and correcting documentation but this is
an improvement, maybe, if only slightly.

The tools that still have to be checked are:

        run_bison
        jsemtblgen
        jnum_gen
        verge
        jsemcgen
        jstr_test
        run_flex
        prep
        mkiocccentry_test
        vermod
        ioccc_test
        hostchk
        jparse_test
        txzchk_test
        utf8_test
        chkentry_test
        reset_tstamp
        bug_report
        jstrdecode
        chkentry
        jstrencode
        jparse
        fnamchk
        iocccsize
        txzchk
        mkiocccentry

but this will probably come another day.