Add clarification for 60760b (or a51c96d1)
I guess the log that was incomplete should have read something like:

    A problem occurred where missing variable declarations in jnum_test.c
    was not fixed properly as after the jnum_test.c has a JNUM_TEST macro
    the jnum_header.c must #include jnum_chk.h. This is because otherwise
    the test_set, test_count and test_result were not previously declared (but
    are in jnum_chk.h). We could not have it static exactly because they're
    used in both source files.

Now as for why I have in parentheses another commit id: it's because I
originally fixed it under fedora and for some reason even though I
pulled from my branch after the changes were pushed, under macOS it
appears to be a different id. I'm not sure if this is because the pull
request has not been merged yet or for some other reason but once it is
I will check and correct the file as I can.