Show $USAGE in scripts when invalid command line
The scripts that did not print the $USAGE string when an invalid
option was specified or when a required argument to an option was not
specified now do so.

At the same time I changed the headers of some scripts adding credits
where credit is due and also (for the jparse specific ones) add that the
'JSON parser was co-developed by ...' in the same way of the files that
do have it.