Plug major memory leaks in jstr{en,de}code.c
The old output of valgrind:

    ==2089136== LEAK SUMMARY:
    ==2089136==    definitely lost: 48 bytes in 1 blocks
    ==2089136==    indirectly lost: 2,245,205 bytes in 2 blocks
    ==2089136==      possibly lost: 0 bytes in 0 blocks
    ==2089136==    still reachable: 0 bytes in 0 blocks
    ==2089136==         suppressed: 0 bytes in 0 blocks

This is because of a typo where == was used instead of != prior to
freeing the buffer as well as not freeing the char *input, both in
jstrencode_stream().

Note however that there still is a minor leak:

    ==2242762== 48 bytes in 1 blocks are definitely lost in loss record 1 of 1
    ==2242762==    at 0x484A464: calloc (vg_replace_malloc.c:1328)
    ==2242762==    by 0x40CDD0: dyn_array_create (dyn_array.c:666)
    ==2242762==    by 0x40AEDB: read_all (util.c:1892)
    ==2242762==    by 0x4028A3: jstrencode_stream (jstrencode.c:88)
    ==2242762==    by 0x40275B: main (jstrencode.c:266)
    ==2242762==
    ==2242762== LEAK SUMMARY:
    ==2242762==    definitely lost: 48 bytes in 1 blocks
    ==2242762==    indirectly lost: 0 bytes in 0 blocks
    ==2242762==      possibly lost: 0 bytes in 0 blocks
    ==2242762==    still reachable: 0 bytes in 0 blocks
    ==2242762==         suppressed: 0 bytes in 0 blocks

Though I could be wrong I believe this is because read_all() uses the
dynamic array facility and although the buffer itself is freed the array
is not.

Whether it's worth fixing this is TBD as it (if this is the reason)
would complicate the code.

For reference: valgrind was used in the jparse/test_jparse/jstr_test.sh
script to detect these leaks.