Exit codes consistency check
Many tools have been checked for exit codes consistency. In some cases
an exit code was reused wrongly. In other cases some exit codes were
added. There were other issues that were fixed as well.

Moved prep.8 from soup/man/man8 to test_ioccc/man/man8 because prep.sh
is under test_ioccc/.

Changed text of the form 'co-developed by:' to be 'co-developed in 2022
by:' as the tools in question (in fact all in the repo) were developed
in 2022.

Changed the exit code of parse_verbosity() to be 3 as that is what most
tools have as a command line error. In one tool checked so far this was
not the case and this was corrected (verge).

The man pages that are yet to be checked for exit codes consistency with
their respective tools are:

    soup/man/man1/chkentry.1
    soup/man/man1/iocccsize.1
    soup/man/man1/mkiocccentry.1
    soup/man/man1/txzchk.1
    soup/man/man8/all_sem_ref.8
    soup/man/man8/reset_tstamp.8
    soup/man/man8/vermod.8
    test_ioccc/man/man1/fnamchk.1
    test_ioccc/man/man8/chkentry_test.8
    test_ioccc/man/man8/hostchk.8
    test_ioccc/man/man8/ioccc_test.8
    test_ioccc/man/man8/ioccc_test.sh.8
    test_ioccc/man/man8/iocccsize_test.8
    test_ioccc/man/man8/mkiocccentry_test.8
    test_ioccc/man/man8/prep.8
    test_ioccc/man/man8/txzchk_test.8

and hopefully this can be done later on today.

In some cases text was reworded a bit and there were some typo fixes as
well.