Add jparse section 3 man pages
Each function documented (the most important ones, for now), have a
symlink to the main jparse.3 and all are in the Makefile so one can
install them as well.

It's likely that more functions could be added and there is no example
but for now it will suffice as more important things have to be done.