Add bison jparse.html report
This seems like useful information to have in the repo as it shows
things like:

    Reductions
    Nonterminals useless in grammar
    Terminals unused in grammar
    Rules useless in grammar
    Conflicts
    Grammar
    Terminals, with rules where they appear
    Nonterminals, with rules where they appear
    Automaton
    States

to show how everything flows along with other information. Some of these
sections are empty like rules useless in grammar but they seem to be
useful information anyway. A note about:

    Terminals unused in grammar

       token

is that this was used as a hack to make better error messages when
unknown tokens are found in input. There might be another way to do it
that we're unaware of but it was one I thought of and it works well so
it's only unused in grammar but not unused.