Improve error messages and exit codes in fnamchk
The error messages needed a bit of work so these were changed. At the
same time I verified that the man page is consistent with the exit codes
and now the usage string also shows the relevant exit codes.

Also an interesting issue with the /*[co]oo*/ in the seqcexit program
that I thought of (and tested and I was right): if the function spans
more than one line then the /*[co]oo*/ line must be on the line with the
function name, not any of the subsequent lines.