More man pages checked for exit codes consistency
This required some changes in exit codes in reset_tstamp.sh. The man
page was updated as well.

The vermod.sh was also checked and both it and the man page were
updated.

The list of man pages that need to be checked for consistency with their
respective tool is now quite short:

    test_ioccc/man/man8/chkentry_test.8
    test_ioccc/man/man8/hostchk.8
    test_ioccc/man/man8/ioccc_test.8
    test_ioccc/man/man8/iocccsize_test.8
    test_ioccc/man/man8/mkiocccentry_test.8
    test_ioccc/man/man8/prep.8
    test_ioccc/man/man8/txzchk_test.8

This I hope to do in the next day or two if not later today.