Check last remaining exit codes for consistency
Some scripts were changed as well as man pages and usage strings.

This commit also includes some improved error messages (including in
mkiocccentry.c).

It also fixes a problem introduced with a change to fnamchk. I thought I
had regenerated the error codes files for txzchk test suite but I must
have not or if I did I changed something again and forgot to do so.

This completes the exit codes consistency checks but there are plenty of
consistency checks in general that have to be done still.