update json_parser_version version
Changed JSON_PARSER_VERSION from "0.12 2022-11-09" to "0.13 2023-01-21".

Added json_parser_version external value containing the library version.

Updated jparse(3) to also mention the dbg_version value.

Fixed how jparse.a is build by NOT adding jparse_main.o to the
library so that the `main()` and `usage()` functions do NOT appear
in the library to avoid confusing other applications that link to
the jparse.a library.

Updated CHANGES.md.