Fix -Wmissing-variable-declarations in sanity.c
With the soup_version string in sanity.c but not in any header file it
triggered a warning that there was no previous extern declaration for
non-static variable:

    sanity.c:22:19: warning: no previous extern declaration for non-static variable 'soup_version' [-Wmissing-variable-declarations]
    const char *const soup_version = SOUP_VERSION;  /* library version format: major.minor YYYY-MM-DD */
		      ^
    sanity.c:22:7: note: declare 'static' if the variable is not intended to be used outside of this translation unit
    const char *const soup_version = SOUP_VERSION;  /* library version format: major.minor YYYY-MM-DD */
	  ^

Now it could have gone in sanity.h but I felt soup.h was more fitting.
If there was a soup.c the version could have gone in there but it seems
like making a soup.c just for that might be overkill - the soup.h
already existed.

Make depend.