Fix bug_report.sh when -x and issues found
In other words if -x was specified if there were any issues found the
log file was still deleted (if any warnings were encountered it was not
deleted). This showed itself in make prep but with other things I was
working on I never got to investigating it (though something looked
wrong).

With this commit if either a warning or an issue is found then the log
file will not be deleted whether -x is specified or not.