Minor adjustments to variable rename (txzchk.[ch])
With commit b5fc2a3658e8eeea711eb908a2fd6c6806b58bae the txzpath was
changed to txzpath_path which seems like it might not be what was
intended. In any case it seems clearer as tarball_path so that's what it
now is called in all cases including the man page and usage message.