Update README.md and jparse/README.md
Added a brief bit of information on how users can help us improve the
repo by testing on other platforms etc. There is room for improvement
here. I noted that we're not yet ready for this. We'll remove that note
when we are ready for it but the how to help information is now there.
In doing this the how to enter was moved to the bottom of the file as
this is not only incomplete at this time but we need it out of the way
for now and rather than delete it I felt moving it below was the easier
way to do it.

Moved the jparse specific files to jparse/README.md and improved that
README file a bit as well.