Add jparse_version string + improve man page
Just like the json_parser_version string we now have one for the jparse
tool. This has been added to the man page.

This change required regenerating the jparse files.

Updated the man page to include the new variable as well as add a
subsection that describes the version strings.