Resolve bug report existing after make prep passes
The problem is that because the make prep only showed errors the
warnings from bug_report.sh were not seen as they're not actually
errors. This meant one would see that make bug_report-txl passed but the
bug report file was still there. This is not a bug but an important
feature of bug_report.sh. This commit improves the prep.sh and
bug_report.sh scripts so that this is more obvious as described below.

Rename 'warning' to 'notice' to not alarm anyone as the checks that
trigger them if they fail are not important.

New option to bug_report.sh: -L logfile which allows for the prep.sh to
specify the log file. This is important so that it can inspect it after
running bug_report.sh. It uses the same format of the bug_report.sh that
is:

    bug-report.$(/bin/date +%Y%m%d.%H%M%S).txt

In bug_report.sh it first sets the bug report to that format and if the
-L option is specified the log file value is changed. It will try
creating a writable log file after parsing options. Although we specify
the same format one could pass manually to prep.sh -L foo.log if they
wanted to. The important point is now prep.sh knows the log file name
which it will, if it still exists (i.e. at least one error or at least
one possible issue detected), grep for the right pattern which is:

    grep -cE "^[[:space:]]+NOTICE:"

Now if no issues are found make prep will look like it used to. If an
error occurs but no notices are triggered it will look like it used to.
If there is at least one error and one issue it might look like:

    [...]
    make_action 24 bug_report-txl -L bug-report.20230124.074635.txt ERROR exit code 48
    make_action 26 seqcexit OK
    make_action 27 picky OK
    make_action 28 tags OK
    make_action 29 check_man OK
    make_action 30 all ERROR exit code 2
    make_action 31 test ERROR exit code 2
    One or more tests failed:

	    make_action 11: make -f ./Makefile all: non-zero exit code: 2
	    make_action 16: make -f ./Makefile all: non-zero exit code: 2
	    make_action 19: make -f ./Makefile all: non-zero exit code: 2
	    make_action 23: make -f ./Makefile all: non-zero exit code: 2
	    make bug_report-txl 24: make bug_report-txl -L bug-report.20230124.074635.txt: non-zero exit code: 48
	    make_action 30: make -f ./Makefile all: non-zero exit code: 2
	    make_action 31: make -f ./Makefile test: non-zero exit code: 2

    One or more tests failed; 4 notices issued.

    See test_ioccc/test_ioccc.log for more details.

    make prep: ERROR: prep.sh exit code: 31

    make prep: see build.log for build details

If only notices are triggered it'll show:

    make_action 24 bug_report-txl -L bug-report.20230124.074757.txt OK
    make_action 25 shellcheck OK
    make_action 26 seqcexit OK
    make_action 27 picky OK
    make_action 28 tags OK
    make_action 29 check_man OK
    make_action 30 all OK
    make_action 31 test OK
    All tests PASSED; 4 notices issued.

    See test_ioccc/test_ioccc.log for more details.

The above assumes make prep which uses the -l option to both prep.sh and
bug_report.sh. It is more verbose when this option is not used.

Updated man page for bug_report.sh.

The above should resolve issue #484.

Symlink prep.sh.8 -> prep.8. This was done because in bug_report.1
prep.sh(8) is referred to but previously only prep.8 existed.