Minor grammar fix in prep.sh
Commit 4bae627277eddb1e5ec68f75f10fbd7b16a23be0 would say even if only
one notice was issued by bug_report.sh that bug_report.sh issued 1
notices which is obviously wrong.

There still is room for improvement in grammar (i.e. some fixes are
almost certainly due) when it comes to plurality in scripts and probably
in code as well but that can come when working on typos and consistency.
This one was fixed as it was inexcusable on my part to have '1 issues'.