improve how bug_report.sh reports on missing commands
The `get_version_optional()` and `get_version()` functions
now try to distinguish between the care where a command
is not found or not executable (and thus the version is unknown) and the
case where the command is found and executable but the
version cannot be determined.

The Notice: messages also distinguish between commands
that are required and those that are optional.