Move static variables from header to source files
There are a few exceptions where they still are in header files. Where
this is the case there is a comment explaining why this is.

There might be some that are missed but many have been moved. This
includes the usage() related strings.

The reason for this change is in the case that another file includes the
header file it might be annoying to have a different version of the
variable.

In the case of txzchk.h probably the txzchk specific variables could
have stayed in txzchk.h but I moved them to txzchk.c anyway.

This commit involved the change of many files but running make prep
works well here.

The possible problematic ones, those under jparse/test_jparse/, have
been tested under both linux and macOS and it passes in both. Since
linux was the one that required some particular changes I believe this
should be good to go.

In general the order of declarations at the top of the source files are
in the same order by type but it's possible that there are some files
that by accident have a different order. If so these will likely be
found later on.

Whether any other variables should be moved is TBD later.