Fix SYNOPSIS and OPTIONS in section 1 man pages
That is to say the options letters (with the '-') are now bold and the
option arguments are italic. I should clarify that the previous commit I
accidentally committed that (for SYNOPSIS; already was there for OPTIONS
section) as I had the change from yesterday but now it uses macros.

Section 8 man pages still need to be updated for this. Section 3 does
not have options and most places macros are used for the SYNOPSIS but
there might be a few places where macros might be used. This is TBD
later.