Fix stray jparse related files after bug_report.sh
After running bug_report.sh two stray jparse related files (without the
apology file prepended) were created in the main directory. It appears
to be related to the old calls:

    ./jparse/run_bison.sh -v 7 -s ./jparse/sorry.tm.ca.h -g ./jparse/verge -D ./jparse
    ./jparse/run_flex.sh -v 7 -s ./jparse/sorry.tm.ca.h -g ./jparse/verge -D ./jparse

This is strange because the scripts would abort if no sorry.tm.ca.h file
was found but then it doesn't appear to be that problem anyway. The
solution was to add two rules to the jparse/Makefile that take the place
of these commands and then, in bug_report.sh, run the rules themselves
rather than the commands directly, via make -C jparse so that the
scripts are run from jparse/ and not the main directory.