Check SYNOPSIS of remaining section 8 man pages
Now all section 8 man pages have had the SYNOPSIS and OPTIONS sections
use macros for the formatting.

This does indeed mean that run_usage.sh is even more useless but it's
still not completely useless. A quick, ugly and dirty hack though it is,
it can still give a starting point for new tools (again with serious
limitations or restrictions) which one can then update the right
formatting.

A word about the man macros that are designed for SYNOPSIS: we don't use
those because they're not guaranteed to be everywhere and there really
is no need for them here either.