cleanup and ignore jparse/lex.yy.c
The `bug_report.sh` tool was creating and leaving behind
`jparse/lex.yy.c` as an untracked file.

This temporary file is now removed by the  `bug_report.sh` tool
near the end of the run, and by the `jparse/Makefile` `make clobber`
rule, and ignored by the .gitignore file.