Typo fix in soup/location.c
Although ions have their place when peroxide ions in particular combine
with water and diluted acids it forms hydrogen peroxide which has ended
up killing people when those unaware of the dangers misunderstand the
request that someone would like a glass of H2O too. Thus we should get
rid of stray ions in the repo and now have.

Yes there's a point (or actually two points) in the above. In order to
help discern what they are you should start by looking at the diff.