Fix formatting for all section 1 man pages
The only case that bold is not enabled/disable by macros is when the
bold is between "s. This might be possible but there are few places
where this occurs and I do not think it's worth spending the time
getting it to work.

What remains on the man page issues to do:

- Check formatting for the other section man pages (3, 8).
- Go through them all (1, 3, 8) to look for consistency problems, typos
etc. (this includes any confusing text or text that might be too
different from the usage string).

After that a final pass can be done. Then once issue #475 is resolved
this can be done for the json semantics tables man pages as well.