Check formatting in all section 3 man pages
There are a few places where the \fB..\fP style pairs are used.

There are probably some minor differences in that in some places we use

    .BR foo,

and in others we use

    .B foo\c
    \&,

and in other places

    .BR foo() \|,

but I think this is not so important. If it should be changed it should
probably be changed to the second or third way but this can be decided
later.

With this commit it means the man pages now need to be checked for
consistencies, typos etc. which hopefully will go much quicker.