Man pages fixes
In the dbg library man pages I changed the .BR foo(), to be

    .BR foo() \|,

like I referred to in commit 3216a670c6f82fcd5e65c105deceb7103dd1c06b.
This was also done in jparse.3 and dyn_array.3.

I also added the version string to the dbg man pages that did not have
it (in other words those that aren't dbg.3).

A few formatting fixes in the dbg man pages as well namely in the
subsection titles (appears to be required to be on all the same line
which I should have known).

The functions in the SEE ALSO section in the dbg(3) man pages are now in
the same order in each page.

Remove stray .fi in jparse.1.