Make names in man pages consistent formatting
There were some places (namely txzchk.1 and txzchk_test.8) where a
certain name was bold (not in all places but in some) namely my name. To
make it more consistent with all the other names (and my name in other
places - except perhaps the README but I would have to check on that
one) I have decided to be less bold. :-)