Fix all typos
This does not mean consistency can be ticked yet. For instance if an
error message refers to some object but the object that it's actually
meant to refer to is not that object (though some have been fixed). It
also does not mean that there aren't some typos I missed. When I go
through files for consistency (if we still want this) I will likely find
a few that slipped through.

I did not check the *.err files or the *.txt files under
test_ioccc/test_txzchk and that includes tar listing output which should
not be touched. The same can be said for the *.json files in the repo.
But everything else should be good unless of course someone changes
things again.

The changes 1st -> first, 2nd -> second and 3rd -> third were redone (I
had previously fixed this but more were added later on).