Check final C file for error messages
With the exception of perhaps dyn_array.c all error messages have been
checked. It is inevitable that I missed some things as it was a lot to
go through and I did not do it by scrolling down from the top line by
line but rather in ways that would help find errors and similar
messages.

In mkiocccentry.c other types of messages were checked and in some cases
improved.

One directory that might need to be rechecked for debug messages is
jparse/ as I did not look for json_dbg() (I did under soup/) but this
can happen another time.

Besides that the only other error (and related) messages to check are
those in shell scripts.