Add FAQ.md, update README.md
As far as the README goes it now links to the FAQ under the new section
'Getting help'. It also links to the FAQ when explaining to report any
issues found with bug_report.sh (under the testing section).

As far as the FAQ goes it has only a few items so far but this can be
considered a work in progress.