Go through remaining json_dbg() calls in jparse/
The following files were modified:

	modified:   jparse/jparse.tab.ref.c
	modified:   jparse/jparse.y
	modified:   jparse/json_parse.c

As far as jparse.y I made numerous json_dbg() calls that were no longer
consistent (due to making the parser re-entrant) consistent. I also
moved a couple calls to json_dbg() to where it matters - namely that if
a certain pointer was NULL the about to perform statements are not
relevant because we would not do that for obvious reasons.

In json_parse.c I moved some json_dbg() calls to the end of the function
as they were above a check for an error which would, if the error was
triggered, make the debug calls not useful or at least not as useful.