Final error consistency checks in C files
I believe that with this commit all C files have been checked for error
consistency issues.

A few error messages were fixed (typos most likely). For instance many
places the word 'is' instead of 'in' were used when the following '%s'
was __func__.

In dyn_test.c parse verbosity failing exited with code 1 but it should
be 3 as it now is.