Improve bug_report.sh
It now tests if each expected subdirectory is searchable, if it has a
Makefile that is readable and if it has an overriding makefile.local
file. These seem like useful things to know even though earlier on we
should be able to tell if the directories have a Makefile.

Some messages were improved a bit as well.