Bug fix hostchk.sh
There were a couple places where it referred to $? but the value was
already overwritten by assigning the new exit code (due to the failure).
For instance if a command exits 1 and you then do:

    EXIT_CODE=42

and then try doing

    echo $?

you'll see 0, not 1.

Rearranged a loop a bit so that the condition is at the top entirely.

Changed some echo messages as well.