Check remaining script files for error messages
With the exceptions noted on GitHub all error messages should be
checked.

I did not check these as thoroughly as in the C files (and I very
possibly missed some there too) but I did discover a number of issues
that were fixed (or else noted on GitHub for discussion).