Further improve FAQ
Add table of contents (numbered) with links.

Although I used 0 for the first number (real programmers start with 0!
:-) ) I know some markdown parsers (or maybe it's a limitation of html
lists?) start at 1. Maybe I should just use 1 for that reason but I
can't help it.

A possible issue is that for each question I have prepended to the
number ## which means it's not an actual list. I'm afraid I don't know
how to do both at the same time both from laziness and loathing anything
to do with web design I have not looked into if this is possible to fix.
Since it's a short list it probably doesn't matter anyway but it seems
important to have them in bigger text. Perhaps I could manually use html
tags but again I really do not like web design. Doing something like:

    0. ##

does seem to work in having both but that appears to make each question
at a column further to the right than the actual answer which is wrong
in my head. Others might look at it differently.