Comment fix in chkentry.c
Remove erroneous statement about what is freed and what is not. Under
linux the streams are closed by the parser and this proves to be a
problem if we try closing it again. That makes sense but why it did not
seem to be a problem under macOS I don't know. In any case we do free
info_path and auth_path but the comments suggested we don't.