Fix make tags
The tags file was not being generated in any directory due to the fact
in the make tags rule the ${CTAGS} had before it a ${V} which is defined
as @: and the : stops the command from running. Now running make tags
will create the tags file in each directory (assuming one has ctags
installed).