Minor updates to party.json files
The bad/party.json file had invalid json. Instead make a new file which
has bad json and let the other one be invalid in semantics.

Some of this will have to be changed when the time comes for the jparse
repo because it's too complex now (and this commit makes it more complex
in a way). I was going to remove some of the files but as we're
essentially at the point of freezing the code I wanted to get this in
quickly and easily.