Improve some of the test scripts
In order to test other implementations of tar the txzchk_test.sh script
has been enhanced to take an optional tar path. To make the tool more
consistent with other tools (except in a way txzchk itself, see below)
the script now is -t tar and -T txzchk. This is because in mkiocccentry
that's the meaning of the options.

The mkiocccentry_test.sh and ioccc_test.sh scripts now also take -t tar
and mkiocccentry_test.sh additionally takes an option for cp (-c cp), ls
(-l ls), txzchk (-T txzchk) and fnamchk (-F fnamchk).

The way these changes are not consistent with txzchk is simply that
txzchk takes a -T option to set text file (or if you prefer 'test') mode
and -t for tar and additionally mkiocccentry takes a -T txzchk so the
scripts that now take a -T that doesn't set txzchk won't match but then
one would hope that txzchk does not need a path to itself! :-)

Updated man pages.