Make hostchk grep usage POSIX compliant
For whatever reason POSIX grep does not require -o or -h options which
we used before. The -o we never actually needed but the -h we do need
because we check more than one file and need only the contents not the
files the contents are in.

Instead of using grep -o -h we pipe the grep to `cut -f 2- -d:' and pipe
that to sort -u (the order is important). This will remove the filenames
from the output and it'll show the exact lines we need but in a POSIX
portable way. This is because (as part of POSIX) if more than one file
is specified to search then the output is prefixed with:

    "%s:", <file>

but we only need the second field in all our cases.

This makes hostchk.sh work on hosts that do not have one or both of grep
-o or grep -h options.